VM: Fix crash in profile signal handler.
The profile signal handler may interrupt when the current Thread
is in an inconsistent state.

Members of the Thread that is unscheduled are NULLed out before
OSThread::SetCurrent sets a new current Thread. Fix this by resetting
the current Thread's fields after a new Thread is set as current.

BUG=#25702
R=iposva@google.com

Review URL: https://codereview.chromium.org/1491283002 .