Revert "Copy entire api-docs directory to release GCloud directory in promote script."
This reverts commit f03279a7eb02ceaac7734c623449e8646cc32664.
When there is only one file in the directory, this does not move the
file into the target directory, but moves it to replace the target
directory with a file.

Reverting to the old behavior, with a changed filename, since we are
now only copying one file. tools/bots/bot_utils.py:124 is changed to
use dartdocs_zipfilename() instead of the deleted apidocs_zipfilename().

BUG=https://github.com/dart-lang/www.dartlang.org/issues/1528
R=ricow@google.com

Review URL: https://codereview.chromium.org/1502253002 .