Fix stack overflow check in InvokeClosure.
Isolate::saved_stack_limit() answers the limit of the Dart stack, which
is different from the C stack when using the simulators. Since we are
recursing in C, we want to check against the C stack's limit.

This overflow check is only needed to support --no-lazy-dispatchers,
which is part of precompilation.

BUG=http://dartbug.com/24659
R=regis@google.com

Review URL: https://codereview.chromium.org/1513993004 .