Make chained futures point to their source instead of opposite.
This means that a chained future that nobody cares about will
not be kept alive by the source.
Any listeners added to the chained future is forwarded to the
source. If the chained source notices that the source has completed,
it copies the values and drops the link completely.

This should allow some unused futures to be GC'ed earlier than
otherwise.

R=floitsch@google.com

Review URL: https://codereview.chromium.org/1516783003.