Merge the prefix and references tables.
This makes the UnlinkedLibrary type go away, which means that we no
longer have any unlinked summary data which is spread across the
compilation units constituting a library.  That in turn means that we
should be able to safely re-link summaries even in the (pathological)
case where a change to URI resolution affects the relationship between a
library and its parts.

R=brianwilkerson@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org/1528113002 .