Indexer one-shot fast path adjustment
If the index/key is not a number it is unlikely that the receiver is an Array-like JavaScript object.

This suggests that we should have a version of the one-shot with no fast-path to use at call sites where we can prove the (receiver, argument) combination does not benefit from the fast path.

BUG=
R=sigmund@google.com

Review URL: https://codereview.chromium.org/1531453004 .