Fix "if" conditions in summary Builder setters.
Previously, we were generating code like:

    if (_value != null || _value == <default>) {
      <store _value in the summary>
    }

which meant that (a) default values were being unnecessarily stored in
the summaries, and (b) passing null to a summary Builder setter would
have caused a crash.

This CL modifies the code generator to output code like this:

    if (!(_value == null || _value == <default>)) {
      <store _value in the summary>
    }

which has the intended behavior of treating `null` like the default
value, and suppressing the default value from appearing in the summary.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1539953002 .