Ignore constructor phase parameter
When --move_super is true (the default value), the super constructor call
is moved to the end of the initializer list. In that case, the phase parameter
is not necessary. This change eliminates the check of the phase parameter in the constructor code, but does not yet eliminate the phase parameter.

BUG=
R=srdjan@google.com

Review URL: https://codereview.chromium.org/1544703002 .