Delay TypeProvider.objectType access until after dart:core resynthesizing.
This way we solve a chicken-and-egg problem with TypeProvider.

Fix for referencing elements from non-first part.

Fix for TODO to update implicit fields for a getter/setter pair.
The type List has it, so I have to fix it to use for SDK summaries.

Add SummaryTypeProvider to use with SDK summaries.
It is initialized in two steps: first dart:core, then dart:async libraries.
    resynthesizer = new SummaryResynthesizer(context, typeProvider,
        _getPrelinkedSummary, _getUnlinkedSummary, context.sourceFactory);
    _buildCoreLibrary();
    _buildAsyncLibrary();

  void _buildCoreLibrary() {
    LibraryElement library = resynthesizer.getLibraryElement('dart:core');
    typeProvider.initializeCore(library);
  }

  void _buildAsyncLibrary() {
    LibraryElement library = resynthesizer.getLibraryElement('dart:async');
    typeProvider.initializeAsync(library);
  }

R=paulberry@google.com, brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/1543313002 .