Add OpenBSD to CheckedInSdkPath
This is used during the build for observatory_tool.py it will attempt
to download an SDK from google but there's none yet for OpenBSD.

This allows the build moving forward with a warning being printed that
this system can't run the SDK and that debug builds will be slow with
the bootstrap dart.

Assuming that upstream will pick up the port this should not be the case
in the future.