Clean up resynthesis of typeArguments.
Previously, we had to jump through hoops to properly resynthesize
InterfaceTypeImpl.typeArguments and FunctionTypeImpl.typeArguments,
because the correct set of type arguments depends on the number of
type parameters, and that information wasn't available in the
prelinked summary.  Now it is.

Note that typeParameters and boundTypeParameters still aren't
resynthesized properly (and are untested).  These will be addressed in
a future CL.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1559123002 .