dart2js cps: Compute intercepted classes in optimize_interceptors.
There was previously an optimization in type propagation that replaced
getInterceptor(x) with x, but this depends on the set of intercepted
classes (since null may or may not need interception).

That optimization is now handled at the use-site of the interceptor,
since each use knows if it needs null to be intercepted.

BUG=
R=kmillikin@google.com, sra@google.com

Review URL: https://codereview.chromium.org/1571433002 .