Support for writing Int32 lists.
We have one List<int> in the summaries model, so we need to be able
to write these lists in FlatBuffers.

We could probably make it more general by adding the Writer interface
and use it for writing any lists, as well as replace specialized
methods like addInt32() or addInt8() with a general addScalarField().
But this will make clients code a little bigger. Thoughts?

R=paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/1568343002 .