dart2js: remove resolved TODO
The answer to the question in the TODO ("does this cover all cases?") is yes.
In Dart, the LHS of a for-in loop must be an identifier. It cannot be a super
access or a static variable access (unless that variable is in scope).

BUG=
R=sra@google.com

Review URL: https://codereview.chromium.org/1580493003 .