Don't suppress compile-time errors when inlining.
When precompiling, it is possible to encounter a compile-time
error when inlining because the target never went through unoptimized
compilation or optimized compilation where it was the root function. Some checks
for compile-time errors are not idemponent because have they have
side-effects: in particular, the checks for a cyclic redirecting factory or
const expressions that have a runtime error may not trigger the second
time they are parsed.

R=fschneider@google.com

Review URL: https://codereview.chromium.org/1568613007 .