Split _LibrarySerializer into two classes.
The new class, _CompilationUnitSerializer, is responsible for
serializing a single compilation unit.  In a future CL, this will allow
us to serialize all the compilation units for a given library in phases,
retaining state information in each of the _CompilationUnitSerializers.

Also, this fixes a previously undiscovered bug where we were trying to
use the same reference index to represent unresolved references across
compilation units; the correct behavior is for each compilation unit to
have its own index for unresolved references.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1601973006 .