Improve 'length' instance property reference encoding.
UnlinkedConstOperation.length is used not only if we can prove that
the target is a StringLiteral (and actually we cannot encode it as a
reference).

The new ReferenceKind.length was added - it is used to encode reference
to the the only instance property we support - 'length'.

R=paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/1642483002 .