dart2js CPS: Remove an invalid assertion in shrinking reductions.
Because we search for new redexes while performing reductions, we can't
always assert that a term is well-formed.

In this case, it's not true that all invocations of a continuation will have
the same number of arguments while we are in the middle of performing a dead
parameter reduction.

BUG=
R=asgerf@google.com

Review URL: https://codereview.chromium.org/1641073003 .