Stop initializing code pages with RWX permissions
We shouldn't need to execute the code pages until after we're done
writing to them, so we can drop the initial execute bit.  Without
this, part of a mapping can sit in a freelist with RWX permissions
indefinitely, which makes for an easier target under certain
exploits.

BUG=
R=asiva@google.com

Review URL: https://codereview.chromium.org/1635343002 .