Use LinkedDependency.parts to find part URIs during resynthesis.
Now we can resynthesize a given library without having to consult any
summaries other than the linked summary for the library itself and the
unlinked summaries for all its compilation units.  All other summaries
are consulted only on demand if the information in them is needed.
Tests are updated to validate this.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1653163003 .