Fix memory leak in incremental resolver
Moving an element that's in the AnalysisCache causes
it to leak because operator== and hashCode are based
on its location.

To make sure we detect this, added a "frozen" flag to
Element. It's set when the element is used as a key
in the AnalysisCache.

Cleanup: in IncrementalResolver, make all fields
final that are never changed.

BUG=
R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1650873002 .