More aggressively fail when associating an existing element model with an AST (issue 25678).
The problem we're seeing is that mismatches between the element model and the AST are causing exceptions to be thrown that are not being handled by the normal exception handling in AnalysisTask. This causes the exception to be converted into an AnalysisException, which will cause the AST structure to be marked as being in ERROR, and it will be re-computed later after a change invalidates it.

This also causes the exception to be thrown more frequently to prevent mismatches from being ignored (and to report the mismatch earlier).

This only checks that every node in the AST has an element associated with it. The next step (in a future CL) is to ensure that the opposite problem doesn't exist (that is, that all elements are associated with a node).

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1669383002 .