Sort members in idl.dart.
I neglected to sort members when converting the idl format from fields
to getters; this rectifies the problem.

Note that sorting members changes the format of summary files, since
data is indexed according to its order of appearance in the idl.  In a
future CL I will introduce a mechanism to keep the indexing stable so
that introducing new fields won't break backward compatibility.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1681733002 .