abort completion request - fixes #24271
This CL aborts completion requests (returns empty list of completions)
if either a newer completion request is received
or the source changes after the completion request was received.

In addition, I wrapped calls to AnalysisFutureHelper computeAsync
in exception handlers so that it will be easier to track async exceptions
back to the code that requested the computations.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1685653002 .