Incrementally update incompletely resolved units.
When the context A uses unit U from the context B, we don't resolve U
in A to RESOLVED_UNIT, we usually stop somewhere before. But we still
want to update U in A incrementally. We don't put RESOLVED_UNITx
results into cache, so we don't write a less resolved unit as a more
resolved result.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/1680403003 .