Ensure that element handles never refer to Member objects.
This is necessary in order for "is" checks (e.g. `element is
ConstructorMember`) to function properly in the presence of element
handles.

To make this possible, it was necessary to lazify the `type` getter in
ExecutableMember, so that attempting to wrap an element handle inside
an ExecutableMember doesn't force the actual element to be
resynthesized.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1686283002 .