CLI. Always fully parse given target files.
It seems that we have a bug - when we analyze a SDK library during
running shared tests, we still parse it dietly so we don't actually
validate that bodies are warning-free.

And when we attempt to use SDK summary, we again parse a library
dietly, but then attempt to apply the full summary using
DeclarationResolver, which fails because we fail to bind some elements
to not parsed nodes.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/1689263004 .