Don't use _inFieldContext tracking.
For any practicaly purpose it was useless because we casted to
FieldDeclaration anyway to get documentation and isStatic field.

One additional change is that we now create ConstFieldElementImpl only
for non-static final fields. This corresponds to the logic in
ResolverVisitor.visitVariableDeclaration(), which clones initializers
only for final non-static fields.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/1696793003 .