Remove redundant verbiage on inference failure errors.
As far as I can tell, the reason is always redundant with the type
error itself. If there are cases where that isn't true, I'd like to
know them.

Also, there was some seemingly dead code in there. _coerceTo() would
return an Identity coercion if fromT <: toT. Then we'd return early from
_checkAssignment().

So the success case in _inferExpression() was never hit: if you get to
the point where it's called, you know it's not a subtype already.

R=leafp@google.com

Review URL: https://codereview.chromium.org/1699463002 .