Fix PEM error checking
While reading PEM data, if we encounter PEM_R_NO_START_LINE, and some
data has been read successfully, we assume that we have reached the
end of the data.

If we encounter PEM_R_NO_START_LINE, and no data has been read
successfully, we assume that the data is not PEM, and we try instead
to read it as PKCS12.

Otherwise, we assume it is malformed PEM, and propagate an error.

R=whesse@google.com

Review URL: https://codereview.chromium.org/1689873003 .