dart2js cps: Do not inline string interpolation helper.
Inlining H.S(x) can lead to code like:
  x === 0 ? 1 / x < 0 ? "-0.0" : "" + x : "" + x

It does not help much to inline this and it can block more useful
inlinings because the extra code brings the caller over the inlining
threshold.

H.S(x) is replaced with x.toString$0(0) if x is a self-interceptor, and
the inliner is then allowed to continue inlining on that.

BUG=
R=kmillikin@google.com, sigmund@google.com

Review URL: https://codereview.chromium.org/1689933002 .