Report a static warning if a for loop condition is non-boolean.
From 17.6.1: "It is a static warning if the static type of c may not be
assigned to bool."

Strong mode catches this case, but it looks like ErrorVerifier just
missed it. Since I'm working on making strong mode lean on ErrorVerifier
for finding sideways casts like this, the discrepancy popped up.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1720083002 .