Propertly summarize LUB function types.
When a function type arises out of a LUB (least upper bound)
computation, it pionts to a synthetic FunctionElement that is not
connected to the element model, so we can't summarize it by simply
storing a reference to the FunctionElement.  We have to store the
FunctionElement's return type and type parameters.

This required reworking _CompilationUnitSerializer's mechanism for
tracking whether it is creating linked or unlinked references;
previously it kept track by passing a boolean around.  Now it uses a
bool field, which should be much less error prone.

Fixes generation of strong mode SDK summary (previously it failed with
an assertion error when in checked mode).

R=scheglov@google.com

Review URL: https://codereview.chromium.org/1740923002 .