dart2js cps: Refactor tracking of side effects.
Primitives now getter for its side effects, as a bitmask.

This avoids a lot of code duplication between GVN, BoundsChecker, and
LoopSideEffects.

There are more flags than those in the SideEffects class (which is used
by the SSA backend, foreign code annotations, and type inference).

We are going to need these extra flags in general, but I want to use
them in the CPS first to see how they work out.

Upcoming changes which are "loose ends" in this CL to keep it manageable:
- Simplify GvnVectorBuilder using dependency flags.
- Stop the overuse of isSafeForElimination.
- Add the extra flags to the rest of the compiler.

BUG=
R=sigmund@google.com

Review URL: https://codereview.chromium.org/1645053002 .